+++ library/src/main/java/com/baasbox/android/json/JsonArray.java
@@ -221,12 +221,7 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
             throw new JsonException(e);
         }
     }
-
-    @Deprecated
-    public JsonArray addBoolean(boolean value){
-        return add(value);
-    }
-    /**
+/**
      * Adds the value at the end of this array
      *
      * @param value
@@ -238,10 +233,7 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
     }
 
 
-    @Deprecated
-    public JsonArray addString(String value){
-        return add(value);
-    }
+
     /**
      * Adds the value at the end of this array
      *
@@ -253,11 +245,6 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
         return this;
     }
 
-
-    @Deprecated
-    public JsonArray addLong(long value){
-        return add(value);
-    }
     /**
      * Adds the value at the end of this array
      *
@@ -270,11 +257,6 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
     }
 
 
-    @Deprecated
-    public JsonArray addDouble(double value){
-        return add(value);
-    }
-
     public JsonArray add(double d) {
         list.add(d);
         return this;
@@ -285,32 +267,14 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
         return this;
     }
 
-
-    @Deprecated
-    public JsonArray addObject(JsonObject value){
-        return add(value);
-    }
-
     public JsonArray add(JsonObject o) {
         list.add(o);
         return this;
     }
 
 
-    @Deprecated
-    public JsonArray addArray(JsonArray value){
-        return add(value);
-    }
-
     public JsonArray add(JsonArray a) {
         list.add(a);
-        return this;
-    }
-
-
-    @Deprecated
-    public JsonArray addBinary(byte[] v) {
-        list.add(v == null ? null : Base64.encode(v, Base64.NO_WRAP));
         return this;
     }
 
@@ -658,21 +622,8 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
         return o;
     }
 
-    @Deprecated
-    public JsonArray setArray(int index, JsonArray value) {
-        list.set(index, value);
-        return this;
-    }
-
 
     public JsonArray set(int index, JsonArray value) {
-        list.set(index, value);
-        return this;
-    }
-
-
-    @Deprecated
-    public JsonArray setBinary(int index, byte[] value) {
         list.set(index, value);
         return this;
     }
@@ -683,11 +634,6 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
     }
 
 
-    @Deprecated
-    public JsonArray setBoolean(int index, boolean value) {
-        list.set(index, value);
-        return this;
-    }
     /**
      * Sets the content at index to the value passed as parameter
      *
@@ -700,21 +646,7 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
         return this;
     }
 
-
-    @Deprecated
-    public JsonArray setDouble(int index, double value) {
-        list.set(index, value);
-        return this;
-    }
-
     public JsonArray set(int index, double value) {
-        list.set(index, value);
-        return this;
-    }
-
-
-    @Deprecated
-    public JsonArray setLong(int index, long value) {
         list.set(index, value);
         return this;
     }
@@ -737,20 +669,7 @@ public class JsonArray extends JsonStructure implements Iterable<Object>, Parcel
     }
 
 
-    @Deprecated
-    public JsonArray setObject(int index, JsonObject value) {
-        list.set(index, value);
-        return this;
-    }
-
     public JsonArray set(int index, JsonObject value) {
-        list.set(index, value);
-        return this;
-    }
-
-
-    @Deprecated
-    public JsonArray setString(int index, String value) {
         list.set(index, value);
         return this;
     }